requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
psycopg2
pandas
pdfplumber
python-multipart
streamlit
plotly
requests
chromadb
onnxruntime
scikit-learn
tiktoken
pymupdf
pymupdf4llm
google-genai
presidio-analyzer
presidio-anonymizer